+++ python_training/midas/engine.py
@@ -1,4 +1,5 @@
-from python_training.config import OPERATIONAL_DB
 from sqlalchemy import create_engine
+
+from python_training.midas.config import OPERATIONAL_DB
 
 engine = create_engine(OPERATIONAL_DB, echo=True)
