+++ python_training/midas/event.py
@@ -1,8 +1,10 @@
-from sqlalchemy import Column, Integer, String, DateTime
-from sqlalchemy.orm import relationship
+from midas
 from python_training.base import Base
 from python_training.event_member_enrollment import event_enrollment_association_table
-from python_training.basic_table import BasicTable
+from sqlalchemy import Column, Integer, String, DateTime
+from sqlalchemy.orm import relationship
+
+from python_training.midas.basic_table import BasicTable
 
 
 class Event(Base, BasicTable):
