+++ python_training/midas/event_member_enrollment.py
@@ -1,5 +1,6 @@
 from sqlalchemy import Table, Column, Integer, ForeignKey
-from python_training.base import Base
+
+from python_training.midas.base import Base
 
 event_enrollment_association_table = Table('event_member_association', Base.metadata,
                                            Column('event_id', Integer,
